Crate penumbra_sdk_ibc

Re-exports

pub use component::ibc_action_with_handler::IbcRelayWithHandlers;

component

pub use component::StateReadExt;

component

pub use component::StateWriteExt;

component

Modules

componentcomponent

genesis

params

Structs

IbcToken

IBC token represents a token that was created through IBC.

Enums

IbcRelay

Statics

IBC_COMMITMENT_PREFIX

The IBC commitment prefix used for the IBC substore, as a MerklePrefix.

IBC_PROOF_SPECS

The ICS23 proof spec for penumbra’s IBC state; this can be used to verify proofs for other substores in the penumbra state, provided that the data is indeed inside a substore (as opposed to directly in the root store.)

IBC_SUBSTORE_PREFIX

The substore prefix used for IBC data.

Traits

MerklePrefixExt

TODO: upstream into ibc-types