penumbra_sdk_ibc::component::packet
Trait SendPacketWrite

pub trait SendPacketWrite: StateWrite {
    // Provided method
    fn send_packet_execute<'life0, 'async_trait>(
        &'life0 mut self,
        packet: IBCPacket<Checked>,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
}

Available on crate feature component only.

This trait, an extension of the Channel, Connection, and Client views, allows a component to send a packet.

Provided Methods

fn send_packet_execute<'life0, 'async_trait>( &'life0 mut self, packet: IBCPacket<Checked>, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Send a packet on a channel. This assumes that send_packet_check has already been called on the provided packet.

Object Safety

This trait is not object safe.